Backend comments added, test file removed

+++ app.py
@@ -146,6 +146,8 @@ def generatesession():
     print(sessionID)
     return jsonify({"success":True, "session":f"{sessionID}"})
 
+
+# Returns true if the session exists and not outdated, False if otherwise
 @app.route("/validatesession", methods=["POST"])
 def validatesession():
     sessionID = request.json.get("sessionID")
@@ -153,6 +155,8 @@ def validatesession():
         return jsonify({"success":False, "message":"invalid session"})
     return jsonify({"success":True})
 
+
+# Returns true if the users roleID is 1, false if its 0
 @app.route("/validateadmin", methods=["POST"])
 def validateadmin():
     sessionID = request.json.get("sessionID")
@@ -162,6 +166,7 @@ def validateadmin():
     return jsonify({"success":True})
 
 
+# Gets the username of a user based on their current sessionID
 @app.route("/getusername", methods=["POST"])
 def getusername():
     session = request.json.get("sessionID")
@@ -171,6 +176,7 @@ def getusername():
     username = db.getData("SELECT username FROM users WHERE userID = (SELECT userID FROM sessions WHERE sessionID = %s)", (session,))[0][0]
     return jsonify({"success":True, "username":username})
 
+# Gets the users role based on their current sessionID
 @app.route("/getrole", methods=["POST"])
 def getrole():
     session = request.json.get("sessionID")
@@ -192,10 +198,11 @@ def updatepassword():
     username = db.getData("SELECT username FROM users WHERE userID = (SELECT userID FROM sessions WHERE sessionID = %s)", (session,))[0][0]
     currentPassword = request.json.get("currentPassword")
 
+    # Checks their password against the value in database
     if not checkPassword(username, currentPassword):
         return jsonify({"success":False, "message":"Incorrect current password"})
     
-    # Takes the new password from the user and updates it in the database
+    # Takes the new password from the user and checks it meets all the requirments for a new password
     newPassword = request.json.get("newPassword")
     if len(newPassword) == 0:
         return jsonify({"success":False, "message":"New password field must not be blank"})
@@ -204,15 +211,16 @@ def updatepassword():
     elif newPassword == currentPassword:
         return jsonify({"success":False, "message":"New password must be different!"})
 
+    # Generates a new salt and hash and updates the database
     salt = db.getData("SELECT salt FROM users WHERE username = %s", (username,))[0][0]
     hashedNewPassword = passwordHashing(newPassword, 20, salt)
     db.manipulateData("UPDATE users SET hashedPassword = %s WHERE username = %s", (hashedNewPassword, username))
     
-
+    # Gets their userID based on their session
     userID = db.getData("SELECT userID FROM sessions WHERE sessionID = %s", (session,))[0][0]
+    
     # Deletes all other sessions
     db.manipulateData("DELETE FROM sessions WHERE userID = %s", (userID,))
-
 
     # Generates a new session with their new password
     sessionID = generateRandomString(64, False)
@@ -222,7 +230,7 @@ def updatepassword():
 
     return jsonify({"success":True, "message":"Password changed", "session":sessionID})
     
-
+# Gets the table names from database
 @app.route("/tables", methods=["GET"])
 def getTables():
     tables = db.getTables()
@@ -232,25 +240,31 @@ def getTables():
 # Route to get data from a specific table
 @app.route('/gettabledata', methods=['POST'])
 def getTableData():
+    # Gets data from the frontned
     data = request.get_json()
+    
+    # Checks for a valid session
     sessionID = data.get('sessionID')  
     if not checkSession(sessionID):
         return jsonify({"success":False, "message":"invalid session"})
 
+    # Checks the user is checking a table that actually exists
     tableName = data.get('tableName')  
     tables = db.getTables()
     excluded_tables = ['sessions', 'users', 'roles']
     tableNames = [table[0] for table in tables if table[0] not in excluded_tables]
     if tableName not in tableNames:
-        print("RAPE")
         return jsonify("")
     
+    # Gets the data from the backend
     query = db.getData(f"SELECT * FROM {tableName}", returnColumnNames=True)
     rowCount = db.getData(f"SELECT COUNT(*) FROM {tableName}")
     return jsonify({"data":query, "rowCount":rowCount})
 
+# Updates the tables from changes from frontend
 @app.route('/updatetables', methods=['POST'])
-def update_table_data():
+def updateTableData():
+    # Primary keys for all tables
     primaryKeyMapping = {
     "categories" : "categoryID",
     "classes" : "classID",
@@ -259,6 +273,7 @@ def update_table_data():
     "suppliers" : "supplierID"
     }
 
+    # Gets data from frontend
     data = request.get_json()  
     session = data['sessionID']
     primaryKeyValue = data['primaryKey']
@@ -267,55 +282,60 @@ def update_table_data():
     tableName = data['tableName']
     primaryKey = primaryKeyMapping[tableName]
 
+    # Checks for a valid session
     if not checkSession(session):
         return jsonify({"success":False, "message":"invalid session"})
     
+    # Checks the table being updated actually exists
     tables = db.getTables()
     excluded_tables = ['sessions', 'users', 'roles']
     tableNames = [table[0] for table in tables if table[0] not in excluded_tables]
     if tableName not in tableNames:
         return jsonify({"success":False})
 
-    requiredRoleID = db.getData(f"SELECT requiredRoleID FROM {tableName} WHERE {primaryKey} = %s", (primaryKeyValue,))[0][0]
-    currentRoleID = db.getData("SELECT roleID FROM users WHERE userID = (SELECT userID FROM sessions WHERE sessionID = %s)", (session,))[0][0]
-
+    # Checks if the value is not NULL and the user is not trying to change the primary key
     if newValue == "":
         return jsonify({"success":False})
-
     if column == primaryKey:
         return jsonify({"success": False, "message":"Cannot change this field"})
     
+    # Checks if the user has permission to edit a field
+    requiredRoleID = db.getData(f"SELECT requiredRoleID FROM {tableName} WHERE {primaryKey} = %s", (primaryKeyValue,))[0][0]
+    currentRoleID = db.getData("SELECT roleID FROM users WHERE userID = (SELECT userID FROM sessions WHERE sessionID = %s)", (session,))[0][0]
     if column == "requiredRoleID" and currentRoleID != 1:
         return jsonify({"success": False, "message":"Cannot change this field"})
-
     if currentRoleID is not None and int(currentRoleID) < int(requiredRoleID):
         return jsonify({"success": False, "message":"Invalid permissions"})
 
+    # Updates the table
     db.manipulateData(f"UPDATE {tableName} SET {column} = %s WHERE {primaryKey} = %s", (newValue, primaryKeyValue))
     return jsonify({"success": True, "message": "Tables updated successfully"}), 200
     
 @app.route('/addrow', methods=['POST'])
-def add_row():
+def addRow():
+    # Gets the data from the frontend
     data = request.get_json()
     sessionID = data.get('sessionID')
     tableName = data.get('tableName')
-
-    tables = db.getTables()
-    excluded_tables = ['sessions', 'users', 'roles']
-    tableNames = [table[0] for table in tables if table[0] not in excluded_tables]
-    if tableName not in tableNames:
-        return jsonify({"success":False})
 
     # Validate session here (you may have a function for this)
     if not checkSession(sessionID):
         return jsonify({"success": False, "message": "Invalid session"})
 
+    # Checks the table they're trying to add a row too exists
+    tables = db.getTables()
+    excluded_tables = ['sessions', 'users', 'roles']
+    tableNames = [table[0] for table in tables if table[0] not in excluded_tables]
+    if tableName not in tableNames:
+        return jsonify({"success":False})
 
+    # Inserts a new row into the table
     db.manipulateData(f"INSERT INTO {tableName} (requiredRoleID) VALUES (0)")
     return jsonify({"success": True, "message": "success"})
 
 @app.route('/deleterow', methods=['POST'])
-def deleterow():
+def deleteRow():
+    # Primary keys for all tables
     primaryKeyMapping = {
     "categories" : "categoryID",
     "classes" : "classID",
@@ -324,22 +344,31 @@ def deleterow():
     "suppliers" : "supplierID"
     }
 
+    # Gets the data from the frontend
     data = request.get_json()
     sessionID = data.get('sessionID')
     tableName = data.get('tableName')
     primaryKeyValue = data.get('primaryKey')
     primaryKey = primaryKeyMapping[tableName]
 
+    # Checks if they have a valid session
     if not checkSession(sessionID):
         return jsonify({"success": False, "message": "Invalid session"})
     
-
+    # Checks they have permission to remove the row
     requiredRoleID = db.getData(f"SELECT requiredRoleID FROM {tableName} WHERE {primaryKey} = %s", (primaryKeyValue,))[0][0]
     currentRoleID = db.getData("SELECT roleID FROM users WHERE userID = (SELECT userID FROM sessions WHERE sessionID = %s)", (sessionID,))[0][0]
-
     if currentRoleID is not None and int(currentRoleID) < int(requiredRoleID):
         return jsonify({"success": False, "message":"Invalid permissions"})
 
+    # Checks the table they're trying to remove a row from exists
+    tables = db.getTables()
+    excluded_tables = ['sessions', 'users', 'roles']
+    tableNames = [table[0] for table in tables if table[0] not in excluded_tables]
+    if tableName not in tableNames:
+        return jsonify({"success":False}) 
+    
+    # Deletes the row from the table
     db.manipulateData(f"DELETE FROM {tableName} WHERE {primaryKey} = %s", (primaryKeyValue,))
     return jsonify({"success": True, "message": "success"})
 
@@ -364,11 +393,12 @@ def assignadmin():
     if not checkUsername(newAdminUsername):
         return jsonify({"success":False, "message":"Username does not exist"})
 
+    # Check if the current users role is an admin
     if checkRole(newAdminUsername) == "Admin":
         return jsonify({"success":False, "message":"User is already an admin"})
 
+    # Updates database to reflect the changes
     db.manipulateData("UPDATE users SET roleID = %s WHERE username = %s", (1, newAdminUsername,))
-    
     return jsonify({"success":True, "message":"Assigned user as admin"})
 
 # Main function to run the program
